Changes to fetch user account details

+++ src/main/java/com/payment/fraud/auth/controller/AdminController.java
@@ -3,7 +3,10 @@ package com.payment.fraud.auth.controller;
 import com.payment.fraud.auth.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
+import com.payment.fraud.auth.entity.User;
 
 import java.util.Set;
 
@@ -24,6 +27,15 @@ public class AdminController {
     public ResponseEntity<String> assignPermission(
             @RequestParam Long userId,
             @RequestParam String permission) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String currentUsername = authentication.getName();
+        User currentUser = userService.findByUsername(currentUsername);
+        boolean isAdmin = currentUser.getRoles().stream()
+                .anyMatch(role -> role.getName().equals("ROLE_ADMIN"));
+        if (!isAdmin) {
+            throw new RuntimeException("Access Denied: Only Admins can assign permissions");
+        }
+
         userService.assignPermission(userId, permission);
         return ResponseEntity.ok("Permission assigned successfully");
     }
@@ -32,7 +44,18 @@ public class AdminController {
     public ResponseEntity<String> removePermission(
             @RequestParam Long userId,
             @RequestParam String permission) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String currentUsername = authentication.getName();
+        User currentUser = userService.findByUsername(currentUsername);
+        boolean isAdmin = currentUser.getRoles().stream()
+                .anyMatch(role -> role.getName().equals("ROLE_ADMIN"));
+
+        if (!isAdmin) {
+            throw new RuntimeException("Access Denied: Only Admins can remove permissions");
+        }
         userService.removePermission(userId, permission);
         return ResponseEntity.ok("Permission removed successfully");
     }
+
+
 }
